+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 * Update development dependencies.
 * Update GraphQL submodule with a fresh schema version.
 * Add logger settings.
+* Add poethepoet as task runner.
 * Add `KUBERNETES` value to `AgentReleasesFilter` enum type.
+* Add `Nrql` to `shortcuts` submodule.
 
 ## [0.23.0] - 2024-11-23